YOLO/detection/requirements.txt
ultralytics
opencv-python
supabase
python-dotenv
requests
